# contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Todo API
  version: 1.0.0
paths:
...
  /api/jira/test:
    post:
      operationId: testJiraConnection
      tags:
        - Jira
      summary: Asks Jira who the stored token belongs to.
      description: >-
        A POST rather than a GET because this is an action the user asks for and expects to
        re-run, not a resource to read; a GET may be served from a cache and answer without
        having contacted Jira at all. /api/jira/statuses stays a GET because a list of status
        names is exactly a cacheable resource.
      responses:
        '200':
          description: The name Jira reports for the token's owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JiraConnectionResponse'
        '400':
          description: Jira could not be reached, or the token was rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/jira/statuses:
    get:
      operationId: listJiraStatuses
      tags:
        - Jira
      summary: Lists the status names the configured project uses.
      responses:
        '200':
          description: The status names, so the waiting list can be picked rather than typed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JiraStatusesResponse'
        '400':
          description: Jira could not be reached, or the token was rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/jira/preview:
    post:
      operationId: previewJira
      tags:
        - Jira
      summary: Reads the user's assigned issues without storing anything.
      responses:
        '200':
          description: The rows the import would create.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JiraPreviewResponse'
        '400':
          description: Jira could not be reached, or the token was rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/jira/import:
    post:
      operationId: importJira
      tags:
        - Jira
      summary: Creates a task for every Jira row that is not imported already.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JiraImportRequest'
      responses:
        '200':
          description: How many rows were created and how many were already there.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JiraImportResponse'
        '400':
          description: The request is not valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    JiraConnectionResponse:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          type: string
    JiraStatusesResponse:
      type: object
      additionalProperties: false
      required: [names]
      properties:
        names:
          type: array
          items:
            type: string
    JiraPreviewRow:
      type: object
      additionalProperties: false
      required: [key, title, url, status, isWaiting, isDuty, alreadyImported, suggestsClosing]
      properties:
        key:
          type: string
        title:
          type: string
        url:
          type: string
          description: >-
            Where Jira shows this issue. Computed from the configured base URL and the key, never
            stored, so it follows a changed base URL. Always present: a preview cannot happen without
            a configured base URL, so this is required rather than nullable — a nullable field would
            add an @if branch, and a branch is unmeasured until a fixture renders it.
        note:
          type: string
          nullable: true
          description: The Jira description, converted from wiki markup to CommonMark.
        deadline:
          type: string
          format: date
          nullable: true
        requester:
          type: string
          nullable: true
        status:
          type: string
          description: The Jira status name, shown so the user can see why a row is waiting.
        isWaiting:
          type: boolean
          description: Whether the status is in the user's waiting list.
        isDuty:
          type: boolean
          description: >-
            Whether this issue came from the duty pool rather than being assigned to the user. The
            screen labels it, so a pool issue is not mistaken for one of your own.
        waitingSince:
          type: string
          format: date-time
          nullable: true
        alreadyImported:
          type: boolean
        suggestsClosing:
          type: boolean
          description: >-
            Whether import should offer to close the local task, because all three of these hold: the
            key has been imported before, the status is in the user's done list, and the local task is
            not done yet. The server's decision rather than three facts for the client to combine —
            the lists live on the server, and the import takes the decision again.
        doneAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            When the issue reached its current status, for a row that suggests closing. Its own field
            rather than a loosened waitingSince, which means something else: a field that changes
            meaning with another boolean is one a later reader misreads. Nullable because Jira's
            changelog may hold no such entry.
        excluded:
          type: string
          nullable: true
          example: jira.excludedWaiting
          description: >-
            Why import will skip this row, as an error code the frontend translates.
            Null means it will be imported.
    JiraPreviewResponse:
      type: object
      additionalProperties: false
      required: [rows, total]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/JiraPreviewRow'
        total:
          type: integer
          format: int32
          description: What Jira reported as the total, so a truncated page is visible.
    JiraImportRow:
      type: object
      additionalProperties: false
      required: [key, title, status]
      properties:
        key:
          type: string
        title:
          type: string
        note:
          type: string
          nullable: true
        deadline:
          type: string
          format: date
          nullable: true
        requester:
          type: string
          nullable: true
        status:
          type: string
          description: >-
            The Jira status name. The server decides whether that means waiting, by looking it
            up in the user's waiting list — the client must not send that decision, because the
            setting lives on the server and a required boolean cannot be enforced on the wire.
        waitingSince:
          type: string
          format: date-time
          nullable: true
    JiraClosureRow:
      type: object
      additionalProperties: false
      required: [key, status]
      properties:
        key:
          type: string
        status:
          type: string
          description: >-
            The Jira status name. Sent for the same reason JiraImportRow sends it: the server decides
            whether that means done by looking it up in the user's done list, and takes the decision
            again rather than trusting a client that may have previewed under an older setting.
        doneAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            What the local task's completion time becomes. A fact the client saw, carried the same way
            waitingSince is — the import deliberately does not call Jira, so it cannot look this up.
    JiraImportRequest:
      type: object
      additionalProperties: false
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/JiraImportRow'
        closures:
          type: array
          items:
            $ref: '#/components/schemas/JiraClosureRow'
          description: >-
            Tasks already imported whose issue is now done, to be closed locally. Optional rather than
            required, so a client that only imports sends what it always sent.
    JiraImportResponse:
      type: object
      additionalProperties: false
      required: [imported, skipped, closed]
      properties:
        imported:
          type: integer
          format: int32
        skipped:
          type: integer
          format: int32
        closed:
          type: integer
          format: int32
          description: How many local tasks were closed because their issue is done.
...